+++ apps/native-component-list/src/screens/ExpoComponentsScreen.tsx
@@ -57,7 +57,9 @@ export const ScreenItems = screens.map((name) => ({
 
 export default function ExpoComponentsScreen() {
   const renderItemRight = React.useCallback(
-    ({ name }) => <ExpoAPIIcon name={name} style={{ marginRight: 10, marginLeft: 6 }} />,
+    ({ name }: { name: string }) => (
+      <ExpoAPIIcon name={name} style={{ marginRight: 10, marginLeft: 6 }} />
+    ),
     []
   );