+++ home/screens/BranchListScreen/BranchListView.tsx
@@ -61,7 +61,7 @@ function BranchList({ data, appId, loadMoreAsync }: Props) {
   const isLoading = React.useRef<null | boolean>(false);
   const theme = useExpoTheme();
 
-  const extractKey = React.useCallback((item) => item.id, []);
+  const extractKey = React.useCallback((item: BranchManifest) => item.id, []);
 
   const handleLoadMoreAsync = async () => {
     if (isLoading.current) return;
@@ -86,18 +86,21 @@ function BranchList({ data, appId, loadMoreAsync }: Props) {
     }
   };
 
-  const renderItem = React.useCallback(({ item: branch, index }) => {
-    return (
-      <BranchListItem
-        key={branch.id}
-        appId={appId}
-        name={branch.name}
-        latestUpdate={branch.latestUpdate}
-        first={index === 0}
-        last={index === data.length - 1}
-      />
-    );
-  }, []);
+  const renderItem = React.useCallback(
+    ({ item: branch, index }: { item: BranchManifest; index: number }) => {
+      return (
+        <BranchListItem
+          key={branch.id}
+          appId={appId}
+          name={branch.name}
+          latestUpdate={branch.latestUpdate}
+          first={index === 0}
+          last={index === data.length - 1}
+        />
+      );
+    },
+    []
+  );
 
   return (
     <View