+++ home/screens/ProjectsListScreen/ProjectList.tsx
@@ -109,7 +109,7 @@ export function ProjectList(props: Props) {
 function ProjectListView({ data, loadMoreAsync }: Props) {
   const isLoading = React.useRef<null | boolean>(false);
   const theme = useExpoTheme();
-  const extractKey = React.useCallback((item) => item.id, []);
+  const extractKey = React.useCallback((item: CommonAppDataFragment) => item.id, []);
 
   const handleLoadMoreAsync = async () => {
     if (isLoading.current) return;