+++ packages/jest-expo/config/getPlatformPreset.js
@@ -37,22 +37,14 @@ function getPlatformPreset(displayOptions, extensions) {
 // Combine React Native for web with React Native
 // Use RNWeb for the testEnvironment
 function getBaseWebPreset() {
-  let reactNativePreset;
-  try {
-    reactNativePreset = require('react-native-web/jest-preset');
-  } catch (error) {
-    console.error(error);
-    throw error;
-  }
   return {
     ...expoPreset,
-    ...reactNativePreset,
-    setupFiles: reactNativePreset.setupFiles,
+    setupFiles: [require.resolve('../src/preset/setup-web.js')],
     moduleNameMapper: {
       ...expoPreset.moduleNameMapper,
       // Add react-native-web alias
       // This makes the tests take ~2x longer
-      ...reactNativePreset.moduleNameMapper,
+      '^react-native$': 'react-native-web',
     },
   };
 }