+++ packages/jest-expo/src/preset/setup-web.js
@@ -1,0 +1,3 @@
+'use strict';
+
+global.__DEV__ = true;